Updated: Lots (see message)
Firstly, this fixes the spacing to be four spaces and not tabs.
Secondly, this adds SIGINT handling on Posix only (Windows coming; this system may not work entirely).
Finally, this adds a new system of handling arguments. We’re now using getopt, allowing for actual arguments and flags. Run the program with the “-h” flag to see it in action.